+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `Neo4J` struct to interact with Neo4j graph database.
 - Add `Chromem` struct to interact with `chromem-go` vector database.
 - Add `Bolt` struct to interact with `bbolt` key-value database.
+- Add `Default` struct to be used by Query and Insert.